♻️ refactor: split §2.4/§2.12 night tests out (ADR-028 slice 4 stage 0)
Pure move of `DesignTokensTests+NightPalette.swift` lines 105-284 into
`DesignTokensTests+NightMeta.swift`. Same per-section axis as stage 0a, and the
axis ADR-028's #1319 Amendment prescribed for slices 3-4 — that file was at 338
lines and slice 4 adds §2.1/§2.3/§2.8 invariants to it.

  DesignTokensTests+NightPalette.swift  338 -> 170
  DesignTokensTests+NightMeta.swift       new 203

No visibility change was needed: slice 3 already widened the three colour-maths
helpers to internal-not-private for its own sibling, so the new file reaches
them as-is.

Verified by running the suite, not by inspection — all seven moved tests execute
from the new file (`nightMetaAndHeaderTokensMatchSpec`,
`nightMetaLadderRunsBrighterFromL1ToL4`,
`nightMetaLadderStaysMonotonicAgainstTheCardSurface`,
`nightLitDotsStayDistinguishableFromUnlitOnes`,
`nightMetaDotOnL1IntentionallyMatchesNightMoss`,
`nightMetaLadderTopIntentionallyMatchesNightInk`,
`headerMetaSubduedReadsTheSameOnBothGrounds`), 80 tests green in one suite.
That is the control against the two ways a moved test file goes silently
un-run — wrong path, or a `-only-testing` zero match that still prints
TEST SUCCEEDED.

Three helper doc comments the move falsified are corrected: `relativeLuminance`
no longer claims "the ordering assertion above" as its purpose, `composite`
points at the file its caller moved to, and the header records the departure.
`contrastRatio` and `composite` now have no caller in their declaring file —
stated explicitly so the next reader does not read them as dead code and follow
the callers.

See #1325
